+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Migrations trigger during module activation
 - Invoice is now downloaded with dynamic filename (invoice number is included)
-- Currency sign shown near prices in the invoice
+- Currency shown near prices in the invoice
+- Footer for invoice document is modifiable in module settings
 
 ## [v1.0.0] - 2022-11-16